Change-detection-due-to-human-activities/requirements.txt
blinker==1.8.2
certifi==2024.7.4
click==8.1.7
cloudinary==1.41.0
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
imageio==2.34.2
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lazy_loader==0.4
MarkupSafe==2.1.5
networkx==3.3
numpy==2.0.1
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
python-dotenv==1.0.1
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
threadpoolctl==3.5.0
tifffile==2024.7.24
urllib3==2.2.2
uuid==1.30
Werkzeug==3.0.3
gunicorn